26 [REFACTO] Logistic Regression module
DONE:
    - Separate utils fonction in submodule
    - Remove HeartDiseaseDataset from LogisticRegressionUtils
    - Train again the model on 10 epochs (2 labels gained accuracy)

TODO:
    - Pull request

issue number : #26
