+++ LogisticRegression/HeartDiseaseDataset.py
@@ -1,8 +1,12 @@
-from torch.utils.data import Dataset, DataLoader
 import numpy as np
-import torch
 import pandas as pd
-from .LogisticRegressionUtils import *
+
+import torch
+from torch.utils.data import Dataset, DataLoader
+
+from .create_dataloaders import create_dataloaders
+from .compute_class_weights import compute_class_weights
+from .compute_min_max import compute_min_max
 
 class HeartDiseaseDataset(Dataset): 
     def __init__(self, path, any_disease=False, label_indexes=[30, 31], split_indexes=[1, 23, 23, 31]):
@@ -16,7 +20,7 @@ class HeartDiseaseDataset(Dataset):
 
         self.minMax = torch.from_numpy(compute_min_max(path, split_indexes[0], split_indexes[1]))
         self.len = len(self.data)
-        self.class_weights = compute_weight_class(self.y)
+        self.class_weights = compute_class_weights(self.y)
 
     def __len__(self):
         return self.len
@@ -24,43 +28,3 @@ class HeartDiseaseDataset(Dataset):
     def __getitem__(self, idx):
         inputs = torch.sub(self.x[idx], self.minMax[0])/ torch.sub(self.minMax[1],self.minMax[0])
         return inputs, self.y[idx]
-
-def create_dataloaders(dataset, batch_size, display_proportions=False):
-    lengths = [round(len(dataset) * split) for split in [TRAIN_SPLIT, VALIDATION_SPLIT, TEST_SPLIT]]
-
-    train_dataset, val_dataset, test_dataset = torch.utils.data.random_split(dataset, lengths=lengths, generator=torch.Generator().manual_seed(SEED))
-
-    train_dataloader = torch.utils.data.DataLoader(
-        train_dataset,
-        batch_size=batch_size,
-        shuffle=True,
-        num_workers=1,
-        prefetch_factor=1,
-        persistent_workers=False,
-        pin_memory=True
-    )
-
-    val_dataloader = torch.utils.data.DataLoader(
-        val_dataset,
-        batch_size=batch_size,
-        shuffle=True,
-        num_workers=4,
-        prefetch_factor=2,
-        persistent_workers=False,
-        pin_memory=True
-    )
-
-    test_dataloader = torch.utils.data.DataLoader(
-        test_dataset,
-        batch_size=1,
-        shuffle=True,
-        num_workers=4,
-        prefetch_factor=2,
-        persistent_workers=False,
-        pin_memory=True
-    )
-
-    if display_proportions:
-        print(f'Total dataset: {len(train_dataloader) + len(val_dataloader) + len(test_dataloader)}, '
-            f'train dataset: {len(train_dataloader)}, val dataset: {len(val_dataloader)}, test_dataset: {len(test_dataloader)}')
-    return train_dataloader, val_dataloader, test_dataloader
