+++ LogisticRegression/compute_class_weights.py
@@ -1,0 +1,7 @@
+import numpy as np
+import torch
+
+def compute_class_weights(labels):
+    unique, counts = np.unique(np.argmax(labels.numpy(), axis=1), return_counts=True)
+    class_weights = torch.tensor([(class_counts / labels.shape[0]) for class_counts in counts],dtype=torch.float32)
+    return class_weights
